requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bitlyshortener==0.6.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==8.0.2
durationpy==0.9
effdet==0.4.1
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
finance-datareader==0.9.96
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
gpt4all==2.8.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.21
langchain-core==0.3.62
langchain-huggingface==0.1.2
langchain-ollama==0.3.1
langchain-openai==0.3.12
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langdetect==1.0.9
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.26
libmagic==1.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
narwhals==1.34.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.71.0
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250327
pi_heif==0.22.0
pikepdf==9.7.0
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pip==25.1.1
platformdirs==4.3.7
playwright==1.51.0
plotly==6.1.2
pluggy==1.6.0
posthog==3.23.0
primp==0.15.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydantic-settings==2.8.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.0.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.1
syrupy==4.9.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.32.3
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0



